requirements.txt
pandas
pdfplumber
pyairtable
python-dotenv
